Specify all dependency versions with less granularity.
This is intended to address a dependency conflict in an app whose other
dependencies require versions of Nap's dependencies greater than the ones
to which it's pinned. The general idea here is to fully trust semantic
versioning rules for post-1.0.0 versions (i.e., assuming that all versions
within a major version number will be backwards-compatible), and going a
bit Wild West on pre-1.0.0 versions (simply setting a minimum version
requirement).